Replacing place_ctx.block_locs std::vector with vtr::vector_map, indexing with ClusterBlockId instead of int

+++ vpr/SRC/base/read_route.cpp
@@ -364,10 +364,10 @@ static void process_global_blocks(ifstream &fp, ClusterNetId inet) {
                         "Block %s for block number %lu specified in the routing file does not match given %s",
                         tokens[1].c_str(), (size_t)bnum, cluster_ctx.clb_nlist.block_name(bnum).c_str());
             }
-            if (place_ctx.block_locs[(size_t)bnum].x != x || place_ctx.block_locs[(size_t)bnum].y != y) {
+            if (place_ctx.block_locs[bnum].x != x || place_ctx.block_locs[bnum].y != y) {
                 vpr_throw(VPR_ERROR_ROUTE, __FILE__, __LINE__,
                         "The placement coordinates (%d, %d) of %d block does not match given (%d, %d)",
-                        x, y, place_ctx.block_locs[(size_t)bnum].x, place_ctx.block_locs[(size_t)bnum].y);
+                        x, y, place_ctx.block_locs[bnum].x, place_ctx.block_locs[bnum].y);
             }
 
 			int pin_index = cluster_ctx.clb_nlist.pin_index(inet, pin_counter);
